feat(widgets): rename /channels to /widgets + catalog review pass
The catalog's URL now matches its nav label and content. /channels
stays as a legacy route: prerendered meta refresh (static hosting has
no server 301s) + canonical to /widgets + instant client replace.
Updated nav, footer, architecture link, sitemap, prerender and
mobile-viewport guards; the /channels API endpoint on /status and the
screenshots/channels asset paths are unrelated and untouched.

Review pass: whole rows are now the ADD TO BAR click target (35 small
buttons made for mean tap targets) with hover tint; the button stays
the accessible aria-pressed control and stops propagation so clicks
never double-fire.
