+++ myscrollr.com/public/sitemap.xml
@@ -43,7 +43,7 @@
     <priority>0.8</priority>
   </url>
   <url>
-    <loc>https://myscrollr.com/channels</loc>
+    <loc>https://myscrollr.com/widgets</loc>
     <lastmod>2026-08-02</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.8</priority>
