+++ myscrollr.com/scripts/generate-sitemap.mjs
@@ -25,7 +25,7 @@ const ROUTES = [
   { path: '/download/linux', priority: 0.8, changefreq: 'weekly' },
   { path: '/uplink', priority: 0.9, changefreq: 'monthly' },
   { path: '/uplink/lifetime', priority: 0.8, changefreq: 'monthly' },
-  { path: '/channels', priority: 0.8, changefreq: 'weekly' },
+  { path: '/widgets', priority: 0.8, changefreq: 'weekly' },
   { path: '/releases', priority: 0.6, changefreq: 'weekly' },
   { path: '/architecture', priority: 0.7, changefreq: 'monthly' },
   { path: '/business', priority: 0.7, changefreq: 'monthly' },
